requirements.txt
pillow
pytesseract
mss
numpy
deep-translator
python-dotenv